fastapi/requirements1.txt
fastapi==0.60.1
pillow==7.1.2
python-multipart==0.0.5 
aiofiles==0.4.0
tensorflow==2.2.0
keras==2.3.1